Add files via upload
I have changed the jumping in Hasan's version to the type in mine. I also got rid of the up command because it didn't make sense for up to move your character when we have a jump key. The only problem is that our two systems don't work together as they are. 

Notice that after a little bit the square falls through the floor. This is because my variable vSpeed is set to increase every tick, and the square is moved every tick by the value of vSpeed. While vSpeed is low, this isn't a problem, the collision with the floor stops the square from moving. Once the value of vSpeed is greater than the thickness of the floor though, the square pops through in one frame. 

I'm thinking the way to fix this is to have a boolean "isGrounded", which is true when the square is colliding with a platform. (Ideally it would only be true when it's on top of a platform if we can do that) Then we can say vSpeed += .03 if (this.isGrounded == false) Do you think this will work?